Fix the issue where HTML snapshot with Unicode-character-named file cannot be manually uploaded
This patch sets the locale inside the container to `C.UTF-8`.

When manually uploading an HTML snapshot, linkding attempts to compress the HTML snapshot into a Gzip file.
Python uses the system's locale settings when encoding binary data. The default locale in the container image is C (POSIX).
When performing encoding operations on HTML snapshot with Unicode characters in the filename, a `UnicodeEncodeError` occurs.